Implement the read-packfile phase of the protocol
This function process the commands received in the previous phase and reads the corresponding pack files. These packfiles will be located in the quarantine folder determined by the GIT_SOCKSTAT_VAR_quarantine_dir environment variable. Please, note, that at this point, if the client has sent a capabilities list it will be read but they are not being used