Fix quarantine building
The quarantine is identified by a quarantine_id in sockstat, and
spokes-receive-pack needs to build the path to the quarantine dir.

Run index-pack with the quarantine env so that it creates the packfile
with a default name but in the quarantine folder. Also set options more
like git-receive-pack.

Include GIT_QUARANTINE_PATH in the quarantine env. git-receive-pack sets
this, so spokes-receive-pack should, too. I'm not sure if this is used
anywhere, but I'm also not sure that it isn't.