Improve the createBogusObjectAndPush pipeline
Use a command stage for hash-object, since it consumes all of its input
and the next stage consumes all of its output.

Change the LinewiseFunction stages into Function stages because they
should run once overall, not once per line. They should only get one
line of input, so these should be equivalent, but this makes it a little
more obvious what the expected behavior is.

Move validation checks (which might call t.Fail) to the main test
goroutine, rather than one of go-pipe's goroutines. This avoids aborting
the other goroutine, which can make go-pipe hang.

Add an EventHandler to the pipeline so that we can see more information
about any errors that occur. In this case, hash-object is failing with
this message:

    error: object fails fsck: missingEmail: invalid author/committer line - missing email

Git 2.40 (which we're now using in CI) is stricter about the input to
hash-object, which is why things started failing. This commit should
change the hang into a test failure that we can debug. The next commit
will fix the new error.