Report the correct git_dir to governor
In prod, the initial working dir is the repository, and the arg list
always says ".". So this isn't really needed there. However, in tests,
the initial working dir is the clone, and the arg list contains the
actual git dir. It is more correct to resolve the argument first,
though, so let's do that.

This change will make it so that certain initialization problems won't
get reported to governor, but we expect them to be vanishingly rare and
those types of problems don't need to be reported to governor.