Handle push-options
Spokes-receive-pack advertises that it can accept push-options so that
pushes to GitHub can use push-options. GitHub doesn't use them, but
custom hooks on GHES might, so we accept them.

Since we advertise it, clients may send push options. If they do, there
will be an extra set of pkt entries between ref updates and the PACK.
This commit adds a step to spokes-receive-pack that consumes the
push-options so that index-pack only sees the PACK.