When sideband is off, send the report directly
Most git clients enable the side-band or side-band-64k capability. When
they do this, the report ("unpack ok" and per-ref status) will be sent
nested in a sideband 1 packet. However, git clients aren't required to
do this, and when they don't the report is supposed to be written
without getting nested. So that's what this change does.