Don't fail is there was some data
When reading the output from spokes-receive-pack, sometimes we get "file
already closed" instead of EOF at the end of the output. In that case,
when there is some data, just log the error and proceed to parsing the
data.

This is an attempt to fix or at least learn more about errors like the
one in https://github.com/github/spokes-receive-pack/actions/runs/4629813916/jobs/8190557968.