Add more data to the die message in governor
The error string provided here will get sent in the "finish" message to
governor. Currently, when index-pack fails, the error message is "exit
code 128". This is a little bit confusing, because it's hard to see just
from the governor output that the problem originates in index-pack. So,
let's also log that the spokes-receive-pack error is due to index-pack's
error so that we know to continue looking in governor for index-pack's
error.