Use os.Chdir less
os.Chdir makes it harder to reason about what's going on, so let's try
to limit how much we use it across the board. In prod code, only do it
once, near the top of Exec. In unit tests, do not use os.Chdir; instead,
set the wd on child processes or pass the dir along. The exceptions are
the test for performReferenceDiscovery, which needs to be able to assume
it's already in the correct directory, and integration tests.