spokes-receive-pack requires a bare repository
git-receive-pack can work with a bare or non-bare repository. However,
we don't need spokes-receive-pack to be able to work on a non-bare
repository.

spokes-receive-pack already will only work on bare repositories, so this
commit isn't a functional change in prod. This change adds a note in
spokes.go that indicates what we're doing. It also changes the one test
that pushed to a non-bare repository (and that is failing now that
connectivity checks are set up correctly) so that it pushes to a bare
repository instead.