index-pack stdout: Take whatever we can get
This should resolve a flaky failure in the governor integration test.

From tests, I've learned that exec.Cmd has a race that sometimes results
in a spurious non-EOF error when reading from a stdout pipe. In this
case, we probably already have the content we need from index-pack's
stdout. In any case, it's safe to pass the bytes along to the channel
because the receiver checks to make sure the data is valid. To firm this
up, I've added a little more validation to the string, which we expect
to be some amount of hex characters.