Create the quarantine dir earlier
One of spokes-receive-pack's jobs is to set up the quarantine directory.
When it doesn't, if another request tries to use the quarantine dir by
setting GIT_OBJECT_DIRECTORY to the quarantine, Git will exit with an
error and say "fatal: not a git repository". CommitQuarantine also
results in an error because the quarantine isn't found. In order to
avoid this error, let's have spokes-receive-pack create the quarantine
dir all the time, not just when there are non-deletes in the push.