.github/workflows: install custom version of git
Future patches are going to add changes which depends on Git features
that are in the `next` branch, but not yet a part of a released version.

Ensure that we can still add features which depend on these unreleased
changes by building and installing Git on `next`.

This patch can safely be reverted after Git 2.42 is released.