internal/spokes/spokes.go: extract reference advertisement format
The following commit is going to potentially invoke `for-each-ref` twice
(once for the bulk of the references, excluding hidden ones, then
optionally again to enumerate any unhidden references).

Our `advertiseRef()` function expects each line of output to match the
format:

    %(objectname) %(refname)

, matching Git's own reference advertisement format.

Extract this portion of the arguments to `for-each-ref` into a constant
to make sure that both invocations produce the same output format.