internal/spokes/spokes.go: avoid enumerating hidden references
To produce the reference advertisement, we run `git for-each-ref` to
produce a listing of all references, and then cull out any references
according to the hidden refs rules.

Usually the list of rules is just to exclude anything under the
`refs/__gh__/` namespace. But for SVN pushes, we unhide anything under
the `refs/__gh__/svn/` namespace. So our `isHidden()` function has to
take these special unhiding rules into account.

Upstream Git has a new feature to exclude certain references via the new
`git for-each-ref --exclude` argument. For the vast majority of pushes,
we can use this feature as-is, since we are never unhiding any
references.

But until we deprecate SVN pushes, we must ensure that we still
advertise anything underneath `refs/__gh__/svn/`, despite hiding the
top-level `refs/__gh__` namespace. To do this, let's use a temporary
workaround to enumerate any unhidden refs as a second pass, and add them
to the list.

Note that this doesn't allow for multi-level hiding/unhiding rules, as
noted in the comment beginning with "NOTE: this assumes that the list of
ref rules is flat [...]". But it should do the trick for our purposes,
and once we finally remove SVN pushes, we can drop the unhiding code
here.