Handle gitmon protocol version
The gitmon protocol will soon be returning the reasons for gitmon
throttling or failing a requests. The gitmon message format changed
to `wait <delay> <reasons>\n` and `fail <delay> <reasons>\n`.

This commit changes the way the gitmon protocol message is currently
parsed, and extract the throttling and failure reasons from it.