Add a package to deal with Git object formats
Git can handle two object formats (hash algorithms) at the current
moment: SHA-1 and SHA-256.  We'd like to add support for SHA-256 to
spokes-receive-pack, so let's add some helpers that let us do so.

Add a type for the object format and also add an accessor that lets us
get the null (all-zeros) object ID as well, since we'll need it
elsewhere.  Expose the constants for these values publicly so we can use
them if need be, and make sure that add some basic tests for our null
OIDs (specifically that they are the proper length).  In addition, make
sure that we can at least properly detect the object format of one of
our test repositories.