spokes: use objectformat null OID constants
We've defined these constants in the objectformat package and exposed
and tested them there.  Replace these inline definitions with references
to the ones in the objectformat package.

The careful reader will note that `nullSHA256OID` has only 60 zeros,
when the real null SHA-256 OID has 64 zeros, since a SHA-256 hash has 64
hex characters.  After this commit, the value is correct.

Note that while we could simply replace references to these existing
variables with actual direct references to the objectformat constants,
that would make the code that uses them a little harder to read and more
verbose, so we don't do that here.