spokes: properly discover and handle object format
We have two places that we need to update the object format in this
package.  First, we need to advertise the correct value in the
`object-format` capability, and second, we need to pass the correct
all-zero (null) OID in the capabilities line.

Let's start by using our new code to look up the object format and then
use that to adjust the capabilities line by moving it into a function
and generating the data appropriately there.

Once we've done that, store the object format into the
`spokesReceivePack` structure and use it to find the null OID
appropriate for the algorithm in use.  This fixes the problem where we'd
always emit a 40-hex character all-zeros object ID for the capabilities.