Allow for custom timeout
The caller should be able to set a custom timeout, and fail closed on
demand.

This change adds a `SCHEDULE_CMD_TIMEOUT` env variable that can be set
by the caller. It also adds a `FAIL_CLOSED` env variable that will abort
the spokes-receive-pack process execution when a timeout was reached,
instead of allowing to run as it was previously the case.