internal/config: introduce `config.ParseSigned()`
Prepare for a change in the next commit which will want to read
integral configuration values like "2G" instead of "2147483648" by
introducing a helper function which knows how to parse such values.

An alternative approach to this patch would be to have Git return these
values to spokes-receive-pack already parsed by calling, for example:

    $ git config --type=int receive.maxSize

where the `--type=int` forces Git to parse the value for us, yielding
the following:

    $ git -c receive.maxSize=2G config --type=int receive.maxSize
    2147483648

But package 'internal/config' gathers all configuration in a single 'git
config --list' invocation, so parsing each as a different type would
require multiple 'git config' invocations, which is unnecessary when the
behavior we're trying to mimmick is relatively straightforward.