internal/spokes: handle suffixed `receive.maxSize`, etc. configuration
When processing an incoming push, spokes-receive-pack assembles the
correct command-line arguments and then invokes 'git index-pack' from
'readPack()' function on the 'spokesReceivePack' struct.

Among the command-line arguments that need to be processed are a handful
of integral values like receive.maxSize, receive.warnObjectSize, and
others. This part of the implementation does roughly:

    maxSize, err := r.getMaxInputSize()
    if err != nil {
            return err
    }

    if maxSize > 0 {
            args = append(args, fmt.Sprintf("--max-input-size=%d", maxSize))
    }

But prior to this commit, the value of the receive.maxSize configuration
was read with strconv.Atoi(), which does not handle suffixed values like
"2G" (which Git interprets as "2 * 1024 * 1024 * 1024").

When spokes-receive-pack encounters a value like this, we return early
from `readPack()`, which causes us to fail all reference update
commands, and return an error from the 'execute' function (also defined
on the spokesReceivePack structure). This causes the user to receive a
push error like the following:

    error: remote unpack failed: index-pack failed

Replace our calls to strconv.Atoi() with the new `config.ParseSigned()`
function defined in the previous commit which correctly handles these
suffixed values, and add an integration test to ensure we prevent
regressions in this area.