internal/spokes: set higher incoming push limit when migrating
For repositories that are being migrated into GitHub, let's set a higher
push limit than our standard 2 GB so that the tool performing that
migration does not have to split up the repository into multiple pushes.

Note that the limit is not infinite here, since we have a hard cap for
the size of an on-disk repository at 100 GB. Setting the limit to 80 GB
ensures that the repository has sufficient room to grow after being
migrated into GitHub.