Allow bad dates, but only for imports and only when it's specifically requested
In general, it's better if customers rewrite their repos to remove the
bad dates because bad dates aren't guaranteed to work with all of our
systems and APIs. But, sometimes there are bad dates in a corner of a
repo where they're unlikely to cause problems, and the cost of rewriting
the repo is too high to be worth it.