Split the is_importing stat in two
We want to use two new different sockstats,
`import_skip_push_limit` and `import_soft_throttling`
which are respectively used to allow a higher push limit,
and implement soft throttling policy for imports.

We still forward the `is_importing` stat for to skip
push limit for backward compatibility. We plan to
remove it on a subsequent PR.